Fixes for warnings.

+++ AutoUpdaterLibrary/test/unit/com/autoupdater/client/download/runnables/TestFileDownloadRunnable.java
@@ -38,6 +38,7 @@ public class TestFileDownloadRunnable extends AbstractTestDownloadRunnable {
                 .exists();
 
         // clean
-        result.delete();
+        if (result != null)
+            result.delete();
     }
 }
